requirements.txt
certifi
setuptools
wheel
aiofiles
aiohttp
aiosignal
albumentations
altair
antlr4-python3-runtime
anyio
async-timeout
attrs
blinker
cachetools
charset-normalizer
click
cmake
contourpy
cycler
decorator
einops
entrypoints
ffmpy
filelock
fonttools
frozenlist
fsspec
ftfy
gitdb
gitpython
gradio
h11
httpcore
httpx
huggingface-hub
idna
imageio
imageio-ffmpeg
importlib-metadata
jinja2
joblib
jsonschema
kiwisolver
lazy-loader
lightning-utilities
linkify-it-py
lit
markdown-it-py
markupsafe
matplotlib
mdit-py-plugins
mdurl
mpmath
multidict
networkx
numpy
omegaconf
open-clip-torch
opencv-contrib-python
opencv-python-headless
orjson
packaging
pandas
pillow
prettytable
protobuf==3.20.3
pyarrow
pydantic
pydeck
pydub
pygments
pympler
pyparsing
pyrsistent
python-dateutil
python-multipart
pytorch-lightning
pytz
pytz-deprecation-shim
pywavelets
pyyaml
qudida
regex
requests
rfc3986
rich
safetensors
scikit-image
scikit-learn
scipy
semver
sentencepiece
six
smmap
sniffio
starlette
streamlit
sympy
threadpoolctl
tifffile
timm
tokenizers
toml
toolz
torchmetrics
tornado
tqdm
transformers
typing-extensions
tzdata
tzlocal==4.3
uc-micro-py
urllib3
uvicorn
validators
watchdog
wcwidth
websockets
yarl
zipp
wandb